backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
email-validator==2.1.0
cryptography==42.0.5
bcrypt==4.1.2
requests==2.31.0
pypdf==3.17.4
pymupdf==1.23.8
numpy==1.24.4
scikit-learn==1.3.2